opschonen
git-svn-id: https://ideas.cs.uu.nl/svn/Feedback/trunk@3709 d9a34ac6-6904-0410-8848-93d71bbb9804

+++ src/Documentation/OverviewPages.hs
@@ -48,7 +48,7 @@ exerciseOverviewPage showAll list = do
          text "all exercises"
       text ", including the ones under development"
       
-   forM_ (zip [1..] (grouping list)) $ \(i, (dom, xs)) -> do
+   forM_ (zip [1::Int ..] (grouping list)) $ \(i, (dom, xs)) -> do
       h2 (show i ++ ". " ++ dom)
       table (map makeRow xs) 
  where
